acn_code/utilities/task_management/tm_to_acn_functions_tested/ai api test/onboarding_testing/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
base58==2.1.1
boto3==1.35.64
botocore==1.35.64
certifi==2024.8.30
colorama==0.4.6
Deprecated==1.2.14
discord.py==2.4.0
distro==1.9.0
ECPy==1.2.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jiter==0.7.0
jmespath==1.0.1
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.53.0
pandas==2.2.3
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.6
types-Deprecated==1.2.9.20240311
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==11.0.3
wrapt==1.16.0
xrpl-py==3.0.0
yarl==1.17.1
